requirements.txt
Django==1.7.7
Genshi==0.7
Markdown==2.6.6
Markups==1.0.1
Pillow==2.3.0
Pygments==2.1.3
Unidecode==0.04.19
Whoosh==2.7.3
amqp==1.4.6
anyjson==0.3.3
argparse==1.1
beautifulsoup4==4.3.2
billiard==3.3.0.18
celery==3.1.23
cssselect==0.9.1
defusedxml==0.4.1
django-allauth==0.19.1
django-autoslug==1.7.2
django-bootstrap3==4.8.2
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.4.0
django-celery==3.1.17
django-crispy-forms==1.4.0
django-dbbackup==2.0.4
-e git+https://github.com/PyAr/django-disqus.git@1dfd32372cf68f77ce2e757ad033abb485d602ae#egg=django_disqus-master
django-email-obfuscator==0.1.3
django-extensions==1.3.3
django-haystack==2.4.1
django-model-utils==2.0.3
django-pagination-py3==1.1.1
django-planet==0.8.3
django-sendfile==0.3.6
django-summernote==0.5.13
django-tagging==0.3.2
django-taggit==0.12.2
django-taggit-autosuggest==0.2.5
docutils==0.12
ecdsa==0.13
feedparser==5.1.3
future==0.15.2
hovercraft==1.1
ipdb==0.8
ipython==3.1.0
kombu==3.0.24
lxml==3.3.5
micawber==0.3.3
oauthlib==1.0.3
paramiko==1.16.0
parse==1.6.6
pdfrw==0.2
pycrypto==2.6.1
pyquery==1.2.11
pysftp==0.2.8
python-telegram-bot==3.4
python3-openid==3.0.9
pytz==2014.4
redis==2.10.1
reportlab==3.3.0
requests==2.2.1
requests-oauthlib==0.6.1
rst2html5==1.7.5
rst2pdf==0.93.dev
sh==1.11
six==1.6.1
svg.path==2.1.1
waliki==0.7dev
